+++ src/ar/edu/unq/po2/alquilerestemporales/ICalificable.java
@@ -3,7 +3,7 @@ package ar.edu.unq.po2.alquilerestemporales;
 import java.util.ArrayList;
 
 public interface ICalificable {
-	public Double getPromedioCalificacion();
+	public float getPromedioCalificacion();
 	public ArrayList<String> getComentarios();
 	public void addCalificacion(Calificacion calificacion);
 }
